fix datepicker array/obj issue
The formElements option passed to DatePicker was incorrectly being
declared as an array ([]) when it should have been as an object ({}).
This works much of the time, but SOMETIMES odd behavior resulted when
iterating through the keys. For instance when attempting to initialize a
datepicker on the events content manager (on Admissions site at
Carleton), the formElements object contained the three DOM elements we
expected to find, but then a bunch of what appear to be JQuery functions
stuffed in there as well! Event content editor on other sites did not
exhibit the bug. Something odd is going on; maybe some area of memory
was getting stale/re-used? Who knows. Anyway properly declaring it as an
object fixes the problem.