remove assumption of organization type from news/post content manager and pub upgrade script
duping a change that was made as a hotfix for 4.7 release. The
publication upgrade script for reaosn 4.7, and the updated news/post
content manager, both assumed that the "organization_type" Reason type
was a core one, when in fact it is Carleton-specific. This caused errors
and odd behavior in the core release.