thor forms send thank you email to the value of "Your Email" first now
Previously, thor forms would only email submitters if the user was logged in.
Now the email goes out to anybody if they supply a valid email address.
It still falls back on netid if the user-supplied email is invalid.

Use case is needed for the Ticket project, where we need to email non-logged
with vital information (i.e. their tickets)