fix a couple thor form bugs regarding tickets addition
one big bug involved some of our payment views that also modify the thor structure, and there was login which pertained to the ticket feature only that was destroying that view's ability to modify the thor xml, too. That's fixed now, but a form with event tickets *and* our standard payment view is likely not compatible out of the box. A couple small tweaks/merges of form info should fix it though.

I think this change also fixed the problem where people delete the last row in the table, but the table wasn't deleted. I think what was happening is that the table was deleted, but then immediately recreated in the same request. This is a smaller bug & I haven't fully verified if its fixed, but these changes seemed to make it go away