tweak CatalogHelper to have course lists match the catalog site
On Department sites when using the Courses page type, the list of courses didn't always match
the list on the Academic Catalog site. There's a feature of these catalog functions that limits/expands
course listings based on the current year *and* site. The site wasn't defined for Department Course pages
since the course list wasn't on the Catalog site. I think this change is actually a bug fix, but there's little documentation
on how to use the Catalog code correctly. Regardless, now course lists show an matching list of related courses
on the Dept website as compared to the current Catalog site.