Docker update (#271)
* Updating docker for dev

Moves things to a more modern docker setup using `docker-compose` and
data/volume containers.

* Adjustments for docker

Now uses the official php 5.6 docker image as a base. Also adds
imagemagick.
MariaDB 5.5 is the database.
Install process is streamlined, uses the docker-install.sh script.
Updated docs.

* Cleanup. Moved docker scripts into the docker folder

* Docker mostly working with docker-compose

Still need to adjust settings to allow for sharing code into the
container to allow for live code updates

* Final fix for docker

docker now has its own copy of the `settings` folder that it copies to
REASON_SETTINGS_PATH on the container when it builds. This allows the
project to be shared into the docker container later so that live code
updates can be reflect in the running container.